Fix static move generation
